Remove redundant projects
