+++ Lighthouse.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26430.16
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lighthouse", "src\Lighthouse\Lighthouse.csproj", "{E945AABA-2779-41E8-9B43-8898FFD64F22}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lighthouse.Tests", "src\Lighthouse.Tests\Lighthouse.Tests.csproj", "{0F9B9BC6-9F86-40E8-BA9B-D27BF3AC7970}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "build", "build", "{183986D1-C1AB-4503-AF06-F2B04A4E51FA}"
@@ -15,9 +13,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "build", "build", "{183986D1
 		build.sh = build.sh
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lighthouse.NetCoreApp", "src\Lighthouse.NetCoreApp\Lighthouse.NetCoreApp.csproj", "{68DA756A-9D3D-4C7E-AA07-8741D2DE0B95}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lighthouse", "src\Lighthouse2\Lighthouse.csproj", "{EFB49E56-91D0-4EA1-BA87-035C863CABE0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lighthouse", "src\Lighthouse\Lighthouse.csproj", "{5CA6B9E7-DAF2-47BD-9C91-7E979B4935A1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,22 +21,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E945AABA-2779-41E8-9B43-8898FFD64F22}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E945AABA-2779-41E8-9B43-8898FFD64F22}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E945AABA-2779-41E8-9B43-8898FFD64F22}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E945AABA-2779-41E8-9B43-8898FFD64F22}.Release|Any CPU.Build.0 = Release|Any CPU
 		{0F9B9BC6-9F86-40E8-BA9B-D27BF3AC7970}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0F9B9BC6-9F86-40E8-BA9B-D27BF3AC7970}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0F9B9BC6-9F86-40E8-BA9B-D27BF3AC7970}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0F9B9BC6-9F86-40E8-BA9B-D27BF3AC7970}.Release|Any CPU.Build.0 = Release|Any CPU
-		{68DA756A-9D3D-4C7E-AA07-8741D2DE0B95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{68DA756A-9D3D-4C7E-AA07-8741D2DE0B95}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{68DA756A-9D3D-4C7E-AA07-8741D2DE0B95}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{68DA756A-9D3D-4C7E-AA07-8741D2DE0B95}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EFB49E56-91D0-4EA1-BA87-035C863CABE0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EFB49E56-91D0-4EA1-BA87-035C863CABE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EFB49E56-91D0-4EA1-BA87-035C863CABE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EFB49E56-91D0-4EA1-BA87-035C863CABE0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5CA6B9E7-DAF2-47BD-9C91-7E979B4935A1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5CA6B9E7-DAF2-47BD-9C91-7E979B4935A1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5CA6B9E7-DAF2-47BD-9C91-7E979B4935A1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5CA6B9E7-DAF2-47BD-9C91-7E979B4935A1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
