+++ README.md
@@ -2,9 +2,9 @@
 
 **Lighthouse** is a simple service-discovery tool for Akka.Cluster, designed to make it easier to place nice with PaaS deployments like Azure / Elastic Beanstalk / AppHarbor.
 
-## Running on .NET 4.5
+## Running on .NET 4.5.2
 
-Lighthouse runs on [Akka.NET](https://github.com/akkadotnet/akka.net) version 1.3.1, which supports .NET 4.5 and .NET Core 1.1/.NET Standard 1.6.  To package the executable and run the .NET 4.5 version locally, clone this repo and build the `Lighthouse` project.  Running Lighthouse.exe in a console should produce an output similar to this:
+Lighthouse runs on [Akka.NET](https://github.com/akkadotnet/akka.net) version 1.3.1, which supports .NET 4.5.2 and .NET Core 1.1/.NET Standard 1.6.  To package the executable and run the .NET 4.5.2 version locally, clone this repo and build the `Lighthouse` project.  Running Lighthouse.exe in a console should produce an output similar to this:
 
 ```
 Topshelf.HostFactory: Configuration Result:
@@ -15,7 +15,7 @@ Topshelf.HostFactory: Configuration Result:
 Topshelf.HostConfigurators.HostConfiguratorImpl: Topshelf v3.2.150.0, .NET Framework v4.0.30319.42000
 ```
 
-The Lighthouse .NET 4.5 project is built as a [Topshelf](https://github.com/Topshelf/Topshelf) service.  This allows you to install Lighthouse as a Windows Service using a command like this:
+The Lighthouse .NET 4.5.2 project is built as a [Topshelf](https://github.com/Topshelf/Topshelf) service.  This allows you to install Lighthouse as a Windows Service using a command like this:
 
 ```
 Lighthouse.exe install --localsystem --autostart
@@ -25,14 +25,14 @@ See the Topshelf documentation for more info on command line arguments for insta
 
 # Running on .NET Core
 
-Lighthouse also includes a .NET Core-compatible version under a separate project named `Lighthouse.NetCoreApp`.  This project does not build as a Topshelf web service.  You have 2 ways that you can run this version:
+Lighthouse also targets the .NET Core 1.1 framework. When building for this target framwork it does not get built as a Topshelf windows service.  You have 2 ways that you can run this version:
 
 - using the .NET CLI
 - building the project as a standalone .exe for your specific [runtime identifier](https://docs.microsoft.com/en-us/dotnet/core/rid-catalog)
 
 #### Using the .NET CLI
 
-Build the project either in Visual Studio 2017 or using `dotnet build -c Release Lighthouse.NetCoreApp.csproj`.  This will output `Lighthouse.NetCoreApp.dll` in your `bin/Release` folder.  From there, running `dotnet ./Lighthouse.NetCoreApp.dll` will start Lighthouse.  Pressing `Enter` will exit.
+Build the project either in Visual Studio 2017 or using `dotnet build -c Release --framework netcoreapp1.1 Lighthouse.csproj`.  This will output `Lighthouse.dll` in your `bin/Release/netcoreapp1.1` folder.  From there, running `dotnet run ./Lighthouse.csproj --framework netcoreapp1.1` will start Lighthouse.  Pressing `Enter` will exit.
 
 #### Building the project as an .exe
 
@@ -56,5 +56,5 @@ bin/
 		netcoreapp1.1/
 			win7-x64/
 				publish/
-					Lighthouse.NetCoreApp.exe
+					Lighthouse.exe
 ```
