--- src/Lighthouse.NetCoreApp/Program.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace Lighthouse.NetCoreApp
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            var lighthouseService = new LighthouseService();
-            lighthouseService.Start();
-            Console.ReadLine();
-            lighthouseService.StopAsync().Wait();
-        }
-    }
-}
