--- src/Lighthouse/App.config
@@ -1,42 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-	<configSections>
-		<section name="akka" type="Akka.Configuration.Hocon.AkkaConfigurationSection, Akka" />
-	</configSections>
-	<akka>
-		<hocon>
-			<![CDATA[
-					lighthouse{
-							actorsystem: "webcrawler" #POPULATE NAME OF YOUR ACTOR SYSTEM HERE
-						}
-			
-					akka {
-						actor { 
-							provider = "Akka.Cluster.ClusterActorRefProvider, Akka.Cluster"
-						}
-						
-						remote {
-							log-remote-lifecycle-events = DEBUG
-							dot-netty.tcp {
-								transport-class = "Akka.Remote.Transport.DotNetty.TcpTransport, Akka.Remote"
-								applied-adapters = []
-								transport-protocol = tcp
-								#will be populated with a dynamic host-name at runtime if left uncommented
-								#public-hostname = "POPULATE STATIC IP HERE"
-								hostname = "0.0.0.0"
-								port = 4053
-							}
-						}            
-						
-						loggers = ["Akka.Logger.NLog.NLogLogger, Akka.Logger.NLog"]
-						
-						cluster {
-							#will inject this node as a self-seed node at run-time
-							seed-nodes = [] #manually populate other seed nodes here, i.e. "akka.tcp://lighthouse@127.0.0.1:4053", "akka.tcp://lighthouse@127.0.0.1:4044"
-							roles = [lighthouse]
-						}
-					}
-			]]>
-		</hocon>
-	</akka>
-</configuration>
