+++ src/Lighthouse/LighthouseHostFactory.cs
@@ -1,20 +1,7 @@
-﻿// Copyright 2014-2015 Aaron Stannard, Petabridge LLC
-//  
-// Licensed under the Apache License, Version 2.0 (the "License"); you may not use 
-// this file except in compliance with the License. You may obtain a copy of the 
-// License at 
-// 
-//     http://www.apache.org/licenses/LICENSE-2.0 
-// 
-// Unless required by applicable law or agreed to in writing, software distributed 
-// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR 
-// CONDITIONS OF ANY KIND, either express or implied. See the License for the 
-// specific language governing permissions and limitations under the License.
-using System.Configuration;
+﻿using System.IO;
 using System.Linq;
 using Akka.Actor;
 using Akka.Configuration;
-using Akka.Configuration.Hocon;
 using ConfigurationException = Akka.Configuration.ConfigurationException;
 
 namespace Lighthouse
@@ -27,8 +14,7 @@ namespace Lighthouse
         public static ActorSystem LaunchLighthouse(string ipAddress = null, int? specifiedPort = null)
         {
             var systemName = "lighthouse";
-            var section = (AkkaConfigurationSection)ConfigurationManager.GetSection("akka");
-            var clusterConfig = section.AkkaConfig;
+            var clusterConfig = ConfigurationFactory.ParseString(File.ReadAllText("akka.hocon"));
 
             var lighthouseConfig = clusterConfig.GetConfig("lighthouse");
             if (lighthouseConfig != null)
@@ -42,7 +28,7 @@ namespace Lighthouse
                         "127.0.0.1"; //localhost as a final default
             int port = specifiedPort ?? remoteConfig.GetInt("dot-netty.tcp.port");
 
-            if(port == 0) throw new ConfigurationException("Need to specify an explicit port for Lighthouse. Found an undefined port or a port value of 0 in App.config.");
+            if (port == 0) throw new ConfigurationException("Need to specify an explicit port for Lighthouse. Found an undefined port or a port value of 0 in App.config.");
 
             var selfAddress = string.Format("akka.tcp://{0}@{1}:{2}", systemName, ipAddress, port);
             var seeds = clusterConfig.GetStringList("akka.cluster.seed-nodes");
