+++ src/Lighthouse/Program.NetCore.cs
