+++ src/Lighthouse/nlog.config
@@ -1,0 +1,13 @@
+<?xml version="1.0" ?>
+<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+    <targets>
+        <target name="eventlog" xsi:type="EventLog" layout="${logger}: ${message} ${exception:format=tostring}" source="Lighthouse" log="Application" />
+        <target name="console" xsi:type="ColoredConsole" layout="${logger}: ${message} ${exception:format=tostring}"></target>
+    </targets>
+
+    <rules>
+        <logger name="*" minlevel="Warn" writeTo="eventlog" />
+        <logger name="*" minlevel="Info" writeTo="console" />
+    </rules>
+</nlog>
