--- src/Lighthouse2/Lighthouse.csproj
@@ -1,27 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFrameworks>netcoreapp1.1;net452</TargetFrameworks>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="Akka" Version="1.3.2" />
-    <PackageReference Include="Akka.Cluster" Version="1.3.2" />
-    <PackageReference Include="Akka.Logger.NLog" Version="1.3.0-beta" />
-    <PackageReference Include="Akka.Remote" Version="1.3.2" />
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="1.1.2" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Xml" Version="1.1.2" />
-  </ItemGroup>
-  <!-- Conditionally obtain references for .NET Framework 4.5.2 -->
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net452' ">
-    <PackageReference Include="Topshelf" Version="4.0.3" />
-    <PackageReference Include="Topshelf.NLog" Version="4.0.3" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Update="akka.hocon">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="NLog.config">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-</Project>
