--- src/Lighthouse2/LighthouseHostFactory.cs
@@ -1,52 +1,0 @@
-﻿using System.IO;
-using System.Linq;
-using Akka.Actor;
-using Akka.Configuration;
-using ConfigurationException = Akka.Configuration.ConfigurationException;
-
-namespace Lighthouse
-{
-    /// <summary>
-    /// Launcher for the Lighthouse <see cref="ActorSystem"/>
-    /// </summary>
-    public static class LighthouseHostFactory
-    {
-        public static ActorSystem LaunchLighthouse(string ipAddress = null, int? specifiedPort = null)
-        {
-            var systemName = "lighthouse";
-            var clusterConfig = ConfigurationFactory.ParseString(File.ReadAllText("akka.hocon"));
-
-            var lighthouseConfig = clusterConfig.GetConfig("lighthouse");
-            if (lighthouseConfig != null)
-            {
-                systemName = lighthouseConfig.GetString("actorsystem", systemName);
-            }
-
-            var remoteConfig = clusterConfig.GetConfig("akka.remote");
-            ipAddress = ipAddress ??
-                        remoteConfig.GetString("dot-netty.tcp.public-hostname") ??
-                        "127.0.0.1"; //localhost as a final default
-            int port = specifiedPort ?? remoteConfig.GetInt("dot-netty.tcp.port");
-
-            if (port == 0) throw new ConfigurationException("Need to specify an explicit port for Lighthouse. Found an undefined port or a port value of 0 in App.config.");
-
-            var selfAddress = string.Format("akka.tcp://{0}@{1}:{2}", systemName, ipAddress, port);
-            var seeds = clusterConfig.GetStringList("akka.cluster.seed-nodes");
-            if (!seeds.Contains(selfAddress))
-            {
-                seeds.Add(selfAddress);
-            }
-
-            var injectedClusterConfigString = seeds.Aggregate("akka.cluster.seed-nodes = [", (current, seed) => current + (@"""" + seed + @""", "));
-            injectedClusterConfigString += "]";
-
-            var finalConfig = ConfigurationFactory.ParseString(
-                string.Format(@"akka.remote.dot-netty.tcp.public-hostname = {0} 
-akka.remote.dot-netty.tcp.port = {1}", ipAddress, port))
-                .WithFallback(ConfigurationFactory.ParseString(injectedClusterConfigString))
-                .WithFallback(clusterConfig);
-
-            return ActorSystem.Create(systemName, finalConfig);
-        }
-    }
-}
